Collapse the classifier back in with the brain.
There's some implementation detail smeared across them that is
not worth untangling at the moment.
