--- src/brain/mod.rs
@@ -1,4 +1,0 @@
-pub mod classifier;
-pub mod loudbot;
-
-pub use loudbot::*;
