+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(non_snake_case)]
 
-pub mod brain;
-pub use brain::*;
-
 pub mod faces;
 pub use faces::*;
+
+pub mod loudbot;
+pub use loudbot::Loudbot;
